fix(类型修正): 将端口类型从size_t更改为int，修正Crypto类中哈希数据函数参数类型，移除handler.h中的不必要包含

+++ AccountingServer.cpp
@@ -78,7 +78,7 @@ int main()
 
     Logger::info(information.str());
     bool ok = server.listen(config.server.host.c_str(), config.server.port); // ctrl+c stop
-    Logger::info("listen return: " + std::to_string(ok));
+    Logger::info("listen return: " + std::to_string(static_cast<int>(ok)));
 
     return 0;
 }
