+++ include/Crypto.h
@@ -19,7 +19,7 @@ public:
 
         CryptAcquireContext(&hProv, nullptr, nullptr, PROV_RSA_AES, CRYPT_VERIFYCONTEXT);
         CryptCreateHash(hProv, CALG_SHA_256, 0, 0, &hHash);
-        CryptHashData(hHash, (BYTE *)_input.c_str(), _input.size(), 0);
+        CryptHashData(hHash, (BYTE *)_input.c_str(), static_cast<DWORD>(_input.size()), 0);
         CryptGetHashParam(hHash, HP_HASHVAL, hash, &hashLen, 0);
 
         CryptDestroyHash(hHash);
